fix some bugs

+++ src/core/dependence_manager.h
@@ -141,7 +141,7 @@ public:
             handle = static_cast<ffrt_task_handle_t>(task);
             outsNoDup.push_back(handle); // handle作为任务的输出signature
         }
-        QoS qos = (attr == nullptr ? QoS() : QoS(static_cast<int>(attr->qos_)));
+        QoS qos = (attr == nullptr ? QoS() : QoS(attr->qos_));
         task->SetQos(qos);
         task->InitRelatedIntervals(parent);
         /* The parent's number of subtasks to be completed increases by one,
